BUG: Ensure order.size is an int
Fixes https://github.com/kernc/backtesting.py/issues/954

+++ backtesting/backtesting.py
@@ -590,7 +590,8 @@ class Trade:
     def close(self, portion: float = 1.):
         """Place new `Order` to close `portion` of the trade at next market price."""
         assert 0 < portion <= 1, "portion must be a fraction between 0 and 1"
-        size = copysign(max(1, round(abs(self.__size) * portion)), -self.__size)
+        # Ensure size is an int to avoid rounding errors on 32-bit OS
+        size = copysign(max(1, int(round(abs(self.__size) * portion))), -self.__size)
         order = Order(self.__broker, size, parent_trade=self, tag=self.__tag)
         self.__broker.orders.insert(0, order)
